Samudrika/requirements.txt
# Samudrika - Underwater Image Enhancement
# PyTorch: install separately for your CUDA version if needed, e.g.:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
